Initial import of gift 0.11.3
What is giFT, you ask? giFT is a modular daemon capable of abstracting the 
communication between the end user and specific filesharing protocols 
(peer-to-peer or otherwise). The giFT project differs from many other similar 
projects in that it is a distribution of a standalone platform-independent 
daemon, a library for client/frontend development, and our own homegrown 
network, OpenFT.

--
http://gift.sourceforge.net/