Initial import of openbox, version 3.0-alpha6:
Openbox is a standards compliant, fast, light-weight, extensible window
manager. 

Openbox works with your applications, and makes your desktop easier to
manage.  This is because the approach to its development was the opposite
of what seems to be the general case for window managers.  Openbox was
written first to comply with standards and to work properly.  Only when
that was in place did the team turn to the visual interface. 

I'm importing this version in -wip because it's a completely rewrite of the
window manager and is still alpha code.  When the final version gets
released, I hope the package in pkgsrc/wm/openbox will be updated to 3.0.