Initial import of afterstep, version 2.00beta1:
AfterStep is a window manager for the Unix X Window System.  Based on the
look and feel of the NeXTStep interface, it provides end users with a
consistent, clean, and elegant desktop.

Some of the distinguishing features of AfterStep compared to other window
managers are its low usage of resources, stability and configurability.

I'm importing this package into -wip because it's still beta quality and
segfaults quite often for me :(