Initial import of cgic, provided in PR pkg/22472 by Douglas Wade Needham.
This is the code written to accompany Thomas Boutell's book "CGI
Programming in C & Perl".  It provides a library of routines for use
in CGI programming, and unlike cgilib-0.5, this library handles the
"multipart/form-data" encoding required to use the FILE input control
type as defined in the HTML specification.