Initial import of thy 0.5.
Package provided by Gergely Nagy <algernon@bonehunter.rulez.org> via PR
pkg/21621.

Thy is an http serveer and it was designed to be as fast and lightweight 
as possible. For speed, she uses sendfile(), and does not fork.

Among the features are name-based virtual hosts, userdir and
directory index support; Range, If-Modified-Since, keep-alive,
on-the-fly compression and of course, CGI, IPv6 and TLS support.