Initial import of map-browse 1.0.
Package provided by Richard Rauch <rkr@olib.org> via PR pkg/22642.

map-browse is a simple, hackish toy that I threw together a few years
ago as a tool to aid me in diagnosing bogus coordinate data in a
(non-graphical) MUD.

The "rooms" in the mud had pathnames, sizes, "zones", and 3D coordinate
data, as well as links to other rooms.  Links could be obvious or non-
obvious.  I collected this subset of data for about 8,000 to 8,500 rooms,
allowing me to analyze bad connections, misplaced rooms, etc., using
a visual exploration.

A 3-button mouse, preferably with scroll-wheel, is strongly recommended.

One small, hand-written sample data file is provided, and an Octave
.m file is also provided which generates 3D data based on the
iteration of a complex function applied at a few points in the complex
plane near 1+0i.

It can be modified, used as a basis perhaps for a game, toyed with
as-is, etc.