Initial import of p5-Scalar-List-Utils 1.11.
Package provided by Koji Kondo <koji@iijmio-mail.jp> via PR pkg/21672.

This package contains a selection of subroutines that people have
expressed would be nice to have in the perl core, but the usage would
not really be high enough to warrant the use of a keyword, and the size
so small such that being individual extensions would be wasteful.

This distribution provides:

  min
  max
  minstr
  maxstr
  sum
  reduce
  reftype
  blessed
  weaken   (5.005_57 and later only)
  isweak   (5.005_57 and later only)
  dualvar
  shuffle