Initial import of p5-ExtUtils-PkgConfig 1.01
The pkg-config program retrieves information about installed libraries,
usually for the purposes of compiling against and linking to them.

ExtUtils::PkgConfig is a very simplistic interface to this utility,
intended for use in the Makefile.PL of perl extensions which bind
libraries that pkg-config knows.  It is really just boilerplate code
that you would've written yourself.