Import hugo 3.0.6 (will be moved to pkgsrc after freeze)
Hugo is one of the leading interactive fiction design systems.  In
addition to being capable of producing leading-edge text adventures,
it also allows a developer to incorporate true-color graphics
display (which are optional for those who lack either the hardware
or inclination to view graphics) as well as sound effects and music.
And it's free.

This is the ncurses port, without support for sound or graphics.