Initial import of boson 0.8
Boson is an OpenGL real-time strategy game, with the feeling of
Command&Conquer(tm) or StarCraft(tm) and OpenGL support.
It is designed to run on Unix computers, and is built on top of the 
libkdegames, kde and Qt libraries.

--
This needs the kdegames3 buildlink, if you don't have this, you will
have to create them with pkgsrc/pkgtools/createbuildlink.

http://boson.eu.org/