import of wxhugo-3.0.06:
Hugo is one of the leading interactive fiction design systems.  In
addition to being capable of producing leading-edge text adventures,
it also allows a developer to incorporate true-color graphics
display (which are optional for those who lack either the hardware
or inclination to view graphics) as well as sound effects and music.
And it's free.

This is the wxGTK port, with support for sound or graphics.  It
does not, however, include game development tools.  For these,
install the package hugo.