Long overdue update to 3.0.
The tarball has been available for quite some time (prior to the
announcement) and I was made aware of this. However, I waited to update
the package because, well, I hate one of the changes that was made to it.

Openbox now follows the XDG base dir spec
(http://www.freedesktop.org/standards/basedir-spec/basedir-spec-0.6.html),
and as a result, we have no easy way to specify a configuration file
directory from a pkgsrc point of view. No more USE_PKGINSTALL or
respecting PKG_SYSCONFDIR.

The user will now have to run obconf or manually copy the example
{rc,menu}.xml files to their own place, and point to it as the XDG spec
says, by setting XDG_CONFIG_DIRS. I added a MESSAGE to reflect this.

If there are any complaints about this, feel free to take it up with the
Openbox author. I'm sure it would also be possible to modify it to behave
as we would like, but I don't think it's worth the effort for a window
manager. All of GNOME will soon be following this spec, so it's probably
time to start thinking about how we want to handle it.

Actual changes in Openbox:
* Add auto-raise functionality with the raiseOnFocus option. It adheres to
  the focusDelay time delay.
* Re-added the focusLast option.
* Renamed the hideTimeout option to hideDelay for consistency.
* Added schemas for the configuration and menu files to the distribution.
  Thanks to Peter Chiocchetti for providing these.
* Added a new gnome-panel-control tool, which allows you to control the
  GNOME 2.4 Panel in key bindings, i.e. show the run dialog, or pop up the
  main menu. This was made necessary in 2.4 when key bindings were removed
  from gnome-panel.
* Added a MoveResize context for mouse bindings. Actions in this context
  will be available during an interactive move or resize of a window.
* Key codes can now be used directly in key bindings instead of using the
  key's name. They can be specified such as 0x40 or C-A-0x5d (that is: in
  hexadecimal, with or without modifiers). This is not normally needed and
  is more difficult to read and maintain, but it is available if you want
  to use keys that don't have names.
* Zillions of bug squashings.