Initial import of perl-5.9.0
TODO:
=====
This is a vile copy of perl58 Makefile and patches, with some changes on version and MAINTAINER. Probably this Makefile needs a cleanup.

I'm working on perl-threads, because:
 $ echo "use threads;" | perl -

always returns -1 :/ and theorically with -Duseithreads="y" this mai be
solved, but doesn't works, probably -Dusethreads?? 

ithreads are internal threads
threads are POSIX threads
oldthreads are to use Thread; code...

I think that the main problem of perl is threads, and this may be solved quickly. If anybody could help on this pkg I will be happy.

pancake