grisbi-0.4.2:
Grisbi is a personnal accounting application for Linux, written
with Gnome and Gtk, and is released under the GPL licence.

Our aim is to provide you with the most simple and intuitive software
for basic use, although it can be very powerful if you spend a
little time on the setup.

Grisbi is an application written by French developpers, so it
perfectly respects French accounting rules. Grisbi can manage
multiple accounts, currencies and users. It manages third party,
expenditure and receipt categories, and also budgetary lines,
financial years, and other informations that make Grisbi adapted
for associations (except those that require double entry accounting).

TODO:

Compiles nicely, but doesn't work :(