Initial import of Firebird relational database management system.
This package does *NOT* work currently and uses non-official DISTFILE.
It is very raw and some work must be put in running it. Please read
my comments in Makefile and TODO before doing anything with it.

As usual, if your changes can make this package work better, feel free
to commit them.