Kyra - A Sprite Engine... from a Slightly Different Point of View
What is a Sprite Engine? A Sprite Engine is the drawing and rendering
component of 2D and quasi-3D games. Examples of this kind of game are
Civilization, Donkey Kong (classic arcade), Pharaoh, Zeus, Warcraft, Diablo,
Frogger, and Pirates!, among many others. It is so called because the
"characters" or "little men" are referred to as "sprites".

Simple and Easy to Use. Kyra has a clean and simple C++ interface. Or at
least as simple as an engine can be. It comes with several examples to get
you started, as well as full documentation for the API and the tool chain.

Fully Featured. It is fully featured, supporting top-down, side, and
isometric rendering. It supports the 'Sprite' as its basic type, but also
supports Tiles and user-drawn Canvases. It can draw to a traditional bitmap
surface, and supports OpenGL hardware acceleration.

Industrial Strength. Kyra has a complete tool chain including a sprite
editor and encoder. It's fast and capable, with specialized code for
rendering and rectangle updates.