Initial import of ap-php5.
TODO: This pkg is done, but will be nice to read php5/TODO and write the
rest of modules inside php5 (like thttpd-php5,roxen-php5, etc.)

This pkg is working properly.


PHP is an HTML-embedded scripting language. It is modular, with
some object-oriented features. Much of its syntax is borrowed from
C, Java and Perl with a couple of unique PHP-specific features
thrown in.  The language is designed to allow web developers to
write dynamically generated pages quickly.

This package contains the dynamically loadable PHP5 Apache module.