Initial import of goldwater
The DGEE is implemented as a Goldwater application. 
Goldwater is a highly scalable distributed middleware 
that promotes the building of applications as a collection 
of separate but logically connected processes. Goldwater 
does all the work of routing messages between these 
processes as though they were mere function calls, even 
if they are distributed across several servers, and 
performs queuing and timeouts as necessary. The DGEE 
therefore inherites Goldwaters scalability and resilience 
along with its management and maintenance features.