Update to 1.18 and patch the bug found by
From: "Bruce J.A. Nourish" <bjan@bjan.net>
Hi,

Package works great - I've been looking for something like it for a
while. One minor problem: the program Makefile hardcodes CC=gcc. The
easiest thing to do is simply remove that line altogether, as $CC is
always passed to the makefile in the environment.


--

There's a 2.0 version but is under development and doesn't builds on *BSD.
I think that this package must in pkgsrc.