Initial import of gnome-clipboard-daemon-1.0.
GNOME Clipboard Daemon is a program that keeps the content of your X
clipboard in memory, so the clipboard will not get lost even after you
close the application from which you copied.  It is a daemon - it has
no GUI. You start it and it will run in the background and Just
Work(tm).